+++ CHANGELOG.md
@@ -7,6 +7,15 @@
 - Added batching to import queue jobs.
 - Added batching to sync queue jobs.
 
+### Fixed
+
+- Fixed a bug in which the expected recipients count could fail in sendouts with very large numbers of contacts.
+
+### Removed
+
+- Removed the `SendoutElement::getPendingRecipients()` method. Use `Campaign::$plugin->sendouts->getPendingRecipients()` instead.
+- Removed the `SendoutElement::getPendingRecipientCount()` method. Use `Campaign::$plugin->sendouts->getPendingRecipients()` instead.
+
 ## 3.0.0-beta.1 - 2024-02-19
 
 > {warning} “Legacy” and “Template” segments are no longer available will be deleted in this update. They should be replaced with regular segments